pyproject.toml
# ref https://github.com/python-poetry/poetry/blob/main/pyproject.toml
[tool.poetry]
name = "bioat"
version = "0.12.15"
description = "bioat, a python package & command line toolkit for Bioinformatics and data science!"
license = "Apache-2.0"
authors = ["Huanan Herman Zhao <hermanzhaozzzz@gmail.com>"]
maintainers = ["Huanan Herman Zhao <hermanzhaozzzz@gmail.com>"]
readme = "README.md"
documentation = "https://bioat.readthedocs.io/en/latest/"
repository = "https://github.com/hermanzhaozzzz/bioat"
homepage = "https://github.com/hermanzhaozzzz/bioat"
keywords = ["bioat", "python", "data-science", "bioinformatics", "bioinformatics-tool"]

[tool.poetry.dependencies]
python = ">=3.10.0"
pandas = ">=1.5"
numpy = ">=1.24.1"
tqdm = ">=4.64.1"
fire = ">=0.5.0"
biopython = ">=1.80"
matplotlib = ">=3.7.1"
tabulate = ">=0.9.0"
statsmodels = ">=0.13.5"
requests = ">=2.31.0"
coloredlogs = ">=15.0.1"
psutil = ">=5.9.8"
openpyxl = ">=3.1.2"
matplotlib-inline = ">=0.1.7"
bamnostic = ">=1.1.10"
beautifulsoup4 = ">=4.12.3"
playwright = ">=1.19.0"
selenium = ">=4.1.3"
scikit-learn = {version = ">=1.0.2", optional = true}
scipy = {version = ">=1.7.3", optional = true}

[tool.poetry.extras]
ai = ["scikit-learn", "scipy"]
bio = ["pybedtools", "pysam"]



[tool.poetry.group.dev.dependencies]
toml = ">=0.10.2"
ruff = ">=0.7.1"  # 替换 black、isort、flake8、mypy 等功能

[tool.ruff]
line-length = 88
extend-include = ["*.ipynb"]
lint.extend-ignore = ["E501"]  # 忽略行长限制，让 `black` 风格兼容
lint.select = ["E", "F", "I", "N", "B", "A"]  # 启用特定规则组
lint.fixable = ["ALL"]  # 启用 Ruff 的全部可自动修复规则

[tool.poetry.group.test.dependencies]
pytest = ">=7.1.3"
pytest-cov = ">=3.0"

[tool.poetry.group.github-actions]
optional = true

[tool.poetry.group.github-actions.dependencies]
pytest-github-actions-annotate-failures = "^0.1.7"

[tool.poetry.scripts]
bioat = "bioat.__main__:main"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
